refactor into same test file
