+++ src/test/java/com/johanekstroem/parking/Controller/CustomerControllerTest.java
@@ -88,29 +88,50 @@ public class CustomerControllerTest {
                 .andExpect(jsonPath("$.id").value("2"));
 
     }
-    
-   /*  @Test
-    void postCustomerCarToCustomerShouldCreateCustomersCar() throws Exception {
-        var customer = new Customer();
-        customer.setFirstName("Elona");
-        customer.setLastName("Muska");
-        customer.setId(1L);
-    
-        var car = new Car();
-        car.setRegistrationNumber("Taslaa005");
-        car.setId(1L);
-    
-        Mockito.when(customerRepo.findById(1L)).thenReturn(Optional.of(customer));
-        Mockito.when(carRepo.save(Mockito.any(Car.class))).thenReturn(car);
-    
-        mockMvc.perform(post("/api/customer/1/car")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(asJsonString(car))
-                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isCreated())
-                .andExpect(jsonPath("$.id").exists());
-    } */
-    
+
+    @Test
+    void getCarsReturnStatus200OK() throws Exception {
+        mockMvc.perform(get("/api/cars"))
+                .andExpect(status().isOk());
+    }
+
+    @Test
+    void getAllCarsShouldReturnListOfCars() throws Exception {
+        Car car = new Car();
+        car.setRegistrationNumber("ABC123");
+
+        Mockito.when(carRepo.findAll()).thenReturn(List.of(car));
+
+        mockMvc.perform(get("/api/cars"))
+                .andExpect(content().contentType(MediaType.APPLICATION_JSON))
+                .andExpect(jsonPath("$[0].registrationNumber").value("ABC123"));
+
+    }
+
+    /*
+     * @Test
+     * void postCustomerCarToCustomerShouldCreateCustomersCar() throws Exception {
+     * var customer = new Customer();
+     * customer.setFirstName("Elona");
+     * customer.setLastName("Muska");
+     * customer.setId(1L);
+     * 
+     * var car = new Car();
+     * car.setRegistrationNumber("Taslaa005");
+     * car.setId(1L);
+     * 
+     * Mockito.when(customerRepo.findById(1L)).thenReturn(Optional.of(customer));
+     * Mockito.when(carRepo.save(Mockito.any(Car.class))).thenReturn(car);
+     * 
+     * mockMvc.perform(post("/api/customer/1/car")
+     * .contentType(MediaType.APPLICATION_JSON)
+     * .content(asJsonString(car))
+     * .accept(MediaType.APPLICATION_JSON))
+     * .andExpect(status().isCreated())
+     * .andExpect(jsonPath("$.id").exists());
+     * }
+     */
+
     public static String asJsonString(final Object obj) {
         try {
             final ObjectMapper mapper = new ObjectMapper();
@@ -120,6 +141,5 @@ public class CustomerControllerTest {
             throw new RuntimeException(e);
         }
     }
-   
 
 }
